# doc-root/news-aggregator/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "API documentation for the News Aggregator application.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/news-aggregator/public/api"
        }
    ],
    "paths": {
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get paginated list of articles",
                "description": "Retrieve a list of articles with optional filters and pagination.",
                "operationId": "20f46a4be21edb0179e4df7acaa45c18",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term for articles",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of articles per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get article by ID",
                "operationId": "90f7485e28819211cd54fc2b866bd33c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch article",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to fetch article"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Article": {
                "description": "Article resource schema",
                "properties": {
                    "id": {
                        "description": "ID of the article",
                        "type": "integer"
                    },
                    "source_name": {
                        "description": "Name of the article's source",
                        "type": "string"
                    },
                    "title": {
                        "description": "Title of the article",
                        "type": "string"
                    },
                    "description": {
                        "description": "Short description of the article",
                        "type": "string"
                    },
                    "content": {
                        "description": "Full content of the article",
                        "type": "string"
                    },
                    "url": {
                        "description": "URL to the article",
                        "type": "string"
                    },
                    "url_to_image": {
                        "description": "URL to the article's image",
                        "type": "string"
                    },
                    "category": {
                        "description": "Category of the article",
                        "type": "string"
                    },
                    "author": {
                        "description": "Author of the article",
                        "type": "string"
                    },
                    "published_at": {
                        "description": "Publication date of the article",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Sanctum Bearer Token Authentication",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}